+++ webapp/climate-action-app/src/app/survey/options.tsx
@@ -1,11 +1,15 @@
+import { CheckboxOption, RadioOption } from "../components/minorComponents";
+
+
 // Referral options for the "Who referred you to this survey?" question
-export const referralOptions = [
+export const referralOptions: RadioOption[] = [
     { value: "family", label: "Family" },
     { value: "friends", label: "Friends" },
-    { value: "supervisorOrCoworker", label: "Supervisor/Coworkers" }
+    { value: "supervisorOrCoworker", label: "Supervisor/Coworkers" },
+    { value: "other", label: "Other: ", textValue: true}
 ];
 
-export const inclinationOptions = [
+export const inclinationOptions: RadioOption[] = [
     { value: "notInclined", label: "Not Inclined" },
     { value: "slightlyInclined", label: "Slightly Inclined" },
     { value: "moderatelyInclined", label: "Moderately Inclined" },
@@ -13,7 +17,7 @@ export const inclinationOptions = [
 ];
 
 // Used for largestImpactChoice, willingToChange, and rankedWillingToChange
-export const carbonFootprintCategories = [
+export const carbonFootprintCategories: RadioOption[] = [
     { value: "home", label: "Home" },
     { value: "electricity", label: "Electricity" },
     { value: "diet", label: "Diet" },
@@ -22,16 +26,16 @@ export const carbonFootprintCategories = [
     { value: "otherConsumption", label: "Other Consumption (Example: buying clothes or furniture)" }
 ];
 
-export const effortToBuyLocalFoodOptions = [
+export const effortToBuyLocalFoodOptions: RadioOption[] = [
     { value: "yes", label: "Yes" },
     { value: "no", label: "No" },
     { value: "occasionally", label: "Occasionally" }
 ];
 
-export const willingToEngageOptions = [
+export const willingToEngageOptions: CheckboxOption[] = [
     { value: "friends", label: "I would engage with friends" },
     { value: "family", label: "I would engage with family" },
     { value: "coworkers", label: "I would engage with coworkers" },
     { value: "otherCommunities", label: "I would engage with other communities" },
-    { value: "notOpen", label: "I would not be open to engaging in a community" }
+    { value: "notOpen", label: "I would not be open to engaging in a community", exclusion_groups: ["notOpen"] }
 ];
